requirements.txt
# MedAI Fusion - Python Dependencies

# Web Framework
Flask==2.3.2
Flask-Session==0.5.0
Flask-CORS==4.0.0

# Machine Learning & Data Science
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.2
xgboost==2.0.0

# Deep Learning
# tensorflow==2.13.0 # Optional: comment out if not using ML models


# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.74

# Explainable AI
shap==0.42.1
lime==0.2.0.1

# Database
SQLAlchemy==2.0.19
pymysql==1.1.0

# Utilities
python-dotenv==1.0.0
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2

# API & Deployment
gunicorn==21.2.0
python-decouple==3.8

# Data Validation
marshmallow==3.20.1
